added printable svg files
This commit adds two new svg files, which takes our A5-paper designed catalog and rotates it so that two of them fit on a standard A4 paper.

The intention is to use inkscape to export a PDF of both of these files, so they can be printed double-sided: spanish on the front; english on the back.